Updates: model_functions.py, params.py, train_test_split.py, main.py, README.md
