+++ DL_logic/params.py
@@ -6,3 +6,6 @@ WORKING_DIRECTORY = os.getcwd()
 # HUGGINGFACE MODEL PATHS
 SARCASM_MODEL_PATH = "helinivan/english-sarcasm-detector"
 FAKE_NEWS_MODEL_PATH = "vikram71198/distilroberta-base-finetuned-fake-news-detection"
+
+# CUDA
+CUDA = False
